split IntBindT into own file.
